add same paaword

+++ src/Components/SignUpForm.jsx
@@ -36,41 +36,51 @@ export const SignUpForm = () => {
 
 
 	const info = async () => {
-		const psw = document.getElementById('loginPsw').value
-		const pswC = document.getElementById('loginPsw').value
-		if (isVerified && name && email && password && passwordConfirmed && privacy && phone && rfc) {
-			console.log('holi')
-			db.collection('preSignUp').doc(email).get()
-				.then((res) => {
-				return res.data()
-				})
-				.then((res) => {
-					console.log(res, email, password);
-					auth.createUserWithEmailAndPassword(email, password)
-						.then(() => {
-							console.log('registrado')
-							db.collection("candidates").doc().set({ name, rfc, phone, email, 'recruiter': res.recruiter, status, cv, interviews, 'vacant': res.vacant })
-							db.collection("users").doc(email).set({ name, email, 'permission': 'applicant' });
-							db.collection('preSignUp').doc(email).delete();
-							history.pushState('./');
-						})
-						.catch(() => {
-							console.log('no registrado :c')
-						})
-				})
-				.catch((err) => {
-				console.log(err, 'no se encontró usuario')
-			})
-		} else {
+		if (password !== passwordConfirmed) {
 			addToast(
-				"Ingresa tu correo y contraseña, y muéstranos que no eres un robot",
+				"Las contraseñas no coinciden",
 				{
 					autoDismiss: true,
 					placement: "top-right",
 					appearance: "warning",
 				}
 			);
+		} else {
+			if (isVerified && name && email && password && passwordConfirmed && privacy && phone && rfc) {
+				console.log('holi')
+				db.collection('preSignUp').doc(email).get()
+					.then((res) => {
+						return res.data()
+					})
+					.then((res) => {
+						console.log(res, email, password);
+						auth.createUserWithEmailAndPassword(email, password)
+							.then(async () => {
+								console.log('registrado')
+								db.collection("candidates").doc().set({ name, rfc, phone, email, 'recruiter': res.recruiter, status, cv, interviews, 'vacant': res.vacant })
+								db.collection("users").doc(email).set({ name, email, 'permission': 'applicant' });
+								db.collection('preSignUp').doc(email).delete();
+								history.pushState('./');
+							})
+							.catch(() => {
+								console.log('no registrado :c')
+							})
+					})
+					.catch((err) => {
+						console.log(err, 'no se encontró usuario')
+					})
+			} else {
+				addToast(
+					"Ingresa tu correo y contraseña, y muéstranos que no eres un robot",
+					{
+						autoDismiss: true,
+						placement: "top-right",
+						appearance: "warning",
+					}
+				);
+			}
 		}
+
 	}
 
 	const [isVerified, setIsVerified] = useState(false);
@@ -102,6 +112,7 @@ export const SignUpForm = () => {
 								type="text"
 								id="signUpName"
 								name="nombre"
+								placeholder="Nombre Completo"
 								onChange={(e) => setName(e.target.value)}
 							/>
 						</div>
@@ -111,6 +122,7 @@ export const SignUpForm = () => {
 								type="text"
 								id="signUpRFC"
 								rfc=""
+								placeholder="RFC"
 								onChange={(e) => setRfc(e.target.value)}
 							/>
 						</div>
@@ -120,6 +132,7 @@ export const SignUpForm = () => {
 								type="text"
 								id="signUpnumber"
 								phone=""
+								placeholder= "Teléfono (10 dígitos)"
 								onChange={(e) => setPhone(e.target.value)}
 							/>
 						</div>
@@ -133,6 +146,7 @@ export const SignUpForm = () => {
 								id="singUpUser"
 								email=""
 								//value={email.email}
+								placeholder= "Crea una contraseña"
 								onChange={(e) => setEmail(e.target.value)}
 							/>
 						</div>
@@ -143,6 +157,7 @@ export const SignUpForm = () => {
 								id="loginPsw"
 								password=""
 								//value={password.password}
+								placeholder= "Crea una contraseña"
 								onChange={(e) => setPassword(e.target.value)}
 							/>
 						</div>
@@ -152,6 +167,7 @@ export const SignUpForm = () => {
 								type="password"
 								passwordConfirmed=""
 								id="loginPswConfirmed"
+								placeholder= "Confirma tu contraseña"
 								onChange={(e) => setPasswordConfirmed(e.target.value)}
 							/>
 						</div>
